Crate nominals

nominals

A crate for formatting nominal indicators in various systems.

This type of formatting can be used for list identifiers when creating ordered lists similar to HTML’s <ol> tag. The crate was written to power the List widget in Cushy.

This crate’s original implementation was inspired by Typst’s implementation of list numbering. However, this crate is a fresh implementation with newly-written implementations and types.

Systems Supported

System	Supported

Arabic (0-9)	✅

Chinese (Simplified)	✅

Chinese (Traditional)	✅

Eastern Arabic	✅

Greek Lowercase	✅

Greek Uppercase	✅

Hebrew	✅

Hiragana Aiueo	✅

Hiragana Iroha	✅

Katakana Aiueo	✅

Katakana Iroha	✅

Korean Jamo	✅

Korean Syllable	✅

Letter Lowercase (a-z)	✅

Letter Upper (A-Z)	✅

Persian	✅

Roman Lowercase	✅

Roman Uppercase	✅

Urdu	✅

armenian

bengali

cambodian

cjk-earthly-branch

cjk-heavenly-stem

devangari

ethiopic-numeric

georgian

gujarati

gurmukhi

japanese-formal

japanese-informal

kannada

lao

malayalam

mongolian

myanmar

oriya

tamil

telegu

thai

tibetan

kannada

All missing systems are listed using their CSS list-style-type identifier. Pull requests are welcome for any and all systems!

no_std support

This crate is no_std compatible, and can operate both with and without alloc. Some systems, such as Chinese, require alloc due to reliance on dependencies that require alloc.

Structs

Chain

A combination of two DigitCollection implementations.

Chinesechinese

A Chinese nominal system.

Decimal

Western ASCII numeric digits.

DigitSet

An ordered set of characters that can be treated as digits.

EasternArabic

Eastern Arabic numeric digits.

GreekLower

Greek lowercase alphabet.

GreekUpper

Greek uppercase alphabet.

HangeulJamo

Korean Hangul Jamo alphabet.

HangeulSyllable

Korean Hangul alphabet as pronounced.

Hebrew

Hebrew numerals.

HexLower

Hexadecimal lower ascii characters.

HexUpper

Hexadecimal uppercase ascii characters.

Hiragana

Japanese Hiaragana Aiueo alphabet.

HiraganaIroha

Japanese Hiaragana Iroha alphabet.

Katakana

Japanese Katakana Aiueo alphabet.

KatakanaIroha

Japanese Katakana Iroha alphabet.

LetterLower

ASCII lowercase characters.

LetterUpper

ASCII uppercase characters.

NominalString

A string that can contain most formatted nominals without a heap allocation.

OneBased

Begins counting at 1 when formatting a nominal identifier.

OutOfMemoryError

No additional memory was able to be allocated.

Persian

Persian numeric digits.

Restrict

Restricts a set of digits to a specific length.

RomanLower

Lowercase Roman numerals.

RomanUpper

Uppercase Roman numerals

Urdu

Urdu numeric digits.

Enums

Error

Error types that can arise from formatting nominals in this crate.

Traits

DigitCollection

An ordered collection of digits that can be used as a NominalSystem.

Nominal

A type that can be formatted with a NominalSystem.

NominalSystem

A system of ordered nominal identifiers.

UnsignedInteger

An unsigned integer type.

UnwrapOrDecimal

Unwraps a result with an Error by formatting the erroring nominal in Decimal.

WithNominal

Converts a result from one error type to an Error<N>.