Struct nominals::DigitSet

pub struct DigitSet<const N: usize, const HAS_ZERO: bool> { /* private fields */ }

An ordered set of characters that can be treated as digits.

Const Generics

N: The number of digits in this set.

HAS_ZERO: When true, the first digit in the set represents 0.

For Decimal, HAS_ZERO is true. This produces the ordering when counting from zero goes:

n	0	1	2	…	9	10	11	12	…

nominal	0	1	2	…	9	10	11	12	…

For LetterUpper, if A were treated as a 0, the first symbol that would appear in the “tens” location would be B:

n	0	1	2	…	25	26	27	28	…

nominal	A	B	C	…	Z	BA	BB	BC	…

This is not the expected behavior, so LetterUpper utilizes DigitSet::zeroless to create a set with HAS_ZERO set to false. This produces the order:

n	0	1	2	…	25	26	27	28	…

nominal	A	B	C	…	Z	AA	AB	AC	…

If 1-based counting is desired, DigitSet::one_based returns an updated set that returns Error::NoZeroSymbol if asked to format 0. It produces this order:

n	0	1	2	3	…	26	27	28	29	…

nominal	err	A	B	C	…	Z	AA	AB	AC	…

Implementations

impl<const N: usize> DigitSet<N, true>

pub const fn new(digits: [char; N]) -> Self

Returns a digit set whose first digit represents the 0 digit.

impl<const N: usize> DigitSet<N, false>

pub const fn zeroless(digits: [char; N]) -> Self

Returns a digit set that does not have a symbol representing a 0 digit.