Trait nominals::Nominal

pub trait Nominal: UnsignedInteger
where
    <Self as TryFrom<usize>>::Error: Debug,
    <Self as TryInto<usize>>::Error: Debug,
{
    // Provided methods
    fn to_nominal<N>(self, system: &N) -> NominalString
       where N: NominalSystem<Self> { ... }
    fn try_to_nominal<N>(self, system: &N) -> Result<NominalString, Error<Self>>
       where N: NominalSystem<Self> + ?Sized { ... }
}

A type that can be formatted with a NominalSystem.

Provided Methods

fn to_nominal<N>(self, system: &N) -> NominalString
where N: NominalSystem<Self>,

Returns self formatted as a nominal identifier using system.

Panics

If N::try_format_nominal() returns an error, this function will panic.

fn try_to_nominal<N>(self, system: &N) -> Result<NominalString, Error<Self>>
where N: NominalSystem<Self> + ?Sized,

Tries to format self as a nominal identifier using system.

Errors

Each nominal system can use its own error type. The crate-level error type is Error, and each variant describes why formatting a nominal may fail.

Object Safety

This trait is not object safe.