Trait nominals::NominalSystem

pub trait NominalSystem<T>
where
    T: Nominal,
    <T as TryFrom<usize>>::Error: Debug,
    <T as TryInto<usize>>::Error: Debug,
{
    // Required method
    fn try_format_nominal(&self, nominal: T) -> Result<NominalString, Error<T>>;

    // Provided method
    fn format_nominal(&self, nominal: T) -> NominalString { ... }
}

A system of ordered nominal identifiers.

Required Methods

fn try_format_nominal(&self, nominal: T) -> Result<NominalString, Error<T>>

Tries to format nominal using this system.

Errors

Each nominal system can use its own error type. The crate-level error type is Error, and each variant describes why formatting a nominal may fail.

Provided Methods

fn format_nominal(&self, nominal: T) -> NominalString

Formats nominal using this system.