Trait nominals::UnwrapOrDecimal

pub trait UnwrapOrDecimal {
    // Required method
    fn unwrap_or_decimal(self) -> NominalString;
}

Unwraps a result with an Error by formatting the erroring nominal in Decimal.

Required Methods

fn unwrap_or_decimal(self) -> NominalString

Returns the nominal string or a nominal string formatted using Decimal.