Trait nominals::DigitCollection

pub trait DigitCollection {
    // Required methods
    fn has_zero_digit(&self) -> bool;
    fn zero_based(&self) -> bool;
    fn len(&self) -> usize;
    fn digit(&self, index: usize) -> char;

    // Provided methods
    fn and<Other>(self, other: Other) -> Chain<Self, Other>
       where Self: Sized,
             Other: DigitCollection { ... }
    fn one_based(self) -> OneBased<Self>
       where Self: Sized { ... }
}

An ordered collection of digits that can be used as a NominalSystem.

Required Methods

fn has_zero_digit(&self) -> bool

Returns true if this collection has a symbol representing 0 at index 0.

fn zero_based(&self) -> bool

Returns true if this collection should start counting at 1 instead of 0. This function is only called if has_zero_digit() returns false.

fn len(&self) -> usize

Returns the number of digits in this collection.

fn digit(&self, index: usize) -> char

Returns the digit at location index.

Panics

This function can panic if index >= self.len().

Provided Methods

fn and<Other>(self, other: Other) -> Chain<Self, Other>
where Self: Sized, Other: DigitCollection,

Chains self and other into a single DigitCollection.

fn one_based(self) -> OneBased<Self>
where Self: Sized,

Returns this collection that indicates counting should start at 1 instead of 0.

This has no effect if the collection has a zero digit.